requirements.txt
Django==5.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
torch==2.1.2
openpyxl==3.1.2
python-dotenv==1.0.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1

# 멀티에이전트 시스템 관련 패키지
crewai>=0.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.6
faiss-cpu>=1.7.4
transformers>=4.35.0
accelerate>=0.20.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=3.20.0 